feat: finish implementing distributed generalized linear model

+++ src/simulations/distributed_generalized_linear_model/message.rs
@@ -13,6 +13,6 @@ impl Message for GlmSumRowsMessage {}
 pub struct GlmConcatMessage {
     pub origin: usize,
     pub r_remote: Mat<f64>,
-    pub iter: i32,
+    pub iter: usize,
 }
 impl Message for GlmConcatMessage {}
